Changed print_terminal_plot to use context rather than fixed numbers.
In other words, instead of always skipping 2 lines at start and 2 lines
at end or whatever, the code looks for the plot command before the graph
and the prompt after the graph specifically.

The code also skips blank lines within a row of graphs.

Hopefully this fixes the occasionally nil/blank plots problem and
chopped-off top line problem for good.  There's still a synchronization
issue leading to the random blank lines, but at least this "should" work
around it.